make php_yar_debug no-inline
